+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Changes
 
   * `Chewy::Query#find` method for finding records by id
+  * `min_score` query option support.
 
 # Version 0.6.0